% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{films}
\alias{films}
\title{Films}
\format{
A data frame with 6 rows and 6 variables:
\describe{
\item{\code{title}}{The title of this film.}
\item{\code{episode_id}}{The episode number of this film.}
\item{\code{opening_crawl}}{The opening crawl text at the beginning of this film.}
\item{\code{director}}{The director of this film.}
\item{\code{producer}}{he producer(s) of this film.}
\item{\code{release_date}}{The release date at original creator country.}
}
}
\usage{
films
}
\description{
Films in the Star Wars movie franchise.
}
\references{
\url{https://swapi.dev/}
}
\keyword{datasets}
