% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{films_people}
\alias{films_people}
\title{People in Films}
\format{
A data frame with 162 rows and 2 variables:
\describe{
\item{\code{title}}{The title of the film.}
\item{\code{character}}{The name of the character who appeared in the film.}
}
}
\usage{
films_people
}
\description{
Links characters (\code{people}) to the \code{films} in which they appear.
}
\references{
\url{https://swapi.dev/}
}
\keyword{datasets}
