% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{films_planets}
\alias{films_planets}
\title{Planets in Films}
\format{
A data frame with 33 rows and 2 variables:
\describe{
\item{\code{title}}{The title of the film.}
\item{\code{planet}}{The name of the planet that appeared in the film.}
}
}
\usage{
films_planets
}
\description{
Links \code{planets} to the \code{films} in which they appear.
}
\references{
\url{https://swapi.dev/}
}
\keyword{datasets}
