% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{films_vehicles}
\alias{films_vehicles}
\title{Vehicles in Films}
\format{
A data frame with 104 rows and 2 variables:
\describe{
\item{\code{title}}{The title of the film.}
\item{\code{vehicle}}{The name of the vehicle that appeared in the film.}
}
}
\usage{
films_vehicles
}
\description{
Links \code{vehicles} to the \code{films} in which they appear
}
\references{
\url{https://swapi.dev/}
}
\keyword{datasets}
