% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pilots}
\alias{pilots}
\title{Pilots}
\format{
A data frame with 43 rows and 2 variables:
\describe{
\item{\code{pilot}}{The name of the person who piloted the vehicle.}
\item{\code{vehicle}}{The name of the vehicle that was piloted.}
}
}
\usage{
pilots
}
\description{
Links \code{people} to the \code{vehicles} they have piloted.
}
\references{
\url{https://swapi.dev/}
}
\keyword{datasets}
