% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planets}
\alias{planets}
\title{Planets}
\format{
A data frame with 59 rows and 9 variables:
\describe{
\item{\code{name}}{The name of this planet.}
\item{\code{rotation_period}}{The number of standard hours it takes for this planet to complete a single rotation on its axis.}
\item{\code{orbital_period}}{The number of standard days it takes for this planet to complete a single orbit of its local star.}
\item{\code{diameter}}{The diameter of this planet in kilometers.}
\item{\code{climate}}{The climate of this planet. Comma-seperated if diverse.}
\item{\code{gravity}}{A number denoting the gravity of this planet. Where 1 is normal.}
\item{\code{terrain}}{The terrain of this planet. Comma-seperated if diverse.}
\item{\code{surface_water}}{The percentage of the planet surface that is naturally occurring water or bodies of water.}
\item{\code{population}}{The average population of sentient beings inhabiting this planet.}
}
}
\usage{
planets
}
\description{
Planets in the Star Wars universe.
}
\references{
\url{https://swapi.dev/}
}
\keyword{datasets}
