% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Species}
\format{
A data frame with 37 rows and 10 variables:
\describe{
\item{\code{name}}{The name of this species.}
\item{\code{classification}}{The classification of this species.}
\item{\code{designation}}{The designation of this species.}
\item{\code{average_height}}{The average height of this person in centimeters.}
\item{\code{skin_colors}}{A comma-seperated string of common skin colors for this species, none if this species does not typically have skin.}
\item{\code{hair_colors}}{A comma-seperated string of common hair colors for this species, none if this species does not typically have hair.}
\item{\code{eye_colors}}{A comma-seperated string of common eye colors for this species, none if this species does not typically have eyes.}
\item{\code{average_lifespan}}{The average lifespan of this species in years.}
\item{\code{homeworld}}{The URL of a planet resource, a planet that this species originates from.}
\item{\code{language}}{The language commonly spoken by this species.}
}
}
\usage{
species
}
\description{
Species within the Star Wars universe.
}
\references{
\url{https://swapi.dev/}
}
\keyword{datasets}
