% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{starwars_db}
\alias{starwars_db}
\alias{starwars_connect}
\alias{starwars_disconnect}
\title{Connect to the Star Wars Database}
\usage{
starwars_connect()

starwars_disconnect(con)
}
\arguments{
\item{con}{A connection to the Star Wars database}
}
\value{
A connection to the Star Wars database
}
\description{
Provides a connection to a DuckDB database of the Star Wars data.
}
\section{Functions}{
\itemize{
\item \code{starwars_connect}: Connect to the DuckDB database

\item \code{starwars_disconnect}: Disconnect from the DuckDB database
}}

