% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vehicles}
\alias{vehicles}
\title{Starships or Vehicles}
\format{
A data frame with 75 rows and 14 variables:
\describe{
\item{\code{name}}{The name of this vehicle. The common name, such as Sand Crawler.}
\item{\code{type}}{The type of the vehicle: starship or vehicle.}
\item{\code{class}}{The class of the vehicle, source from \code{starship_class} or \code{vehicle_class}.}
\item{\code{model}}{The model or official name of this vehicle. Such as All Terrain Attack Transport.}
\item{\code{manufacturer}}{The manufacturer of this vehicle. Comma seperated if more than one.}
\item{\code{cost_in_credits}}{The cost of this vehicle new, in galactic credits.}
\item{\code{length}}{The length of this vehicle in meters.}
\item{\code{max_atmosphering_speed}}{The maximum speed of this vehicle in atmosphere.}
\item{\code{crew}}{The number of personnel needed to run or pilot this vehicle.}
\item{\code{passengers}}{The number of non-essential people this vehicle can transport.}
\item{\code{cargo_capacity}}{The maximum number of kilograms that this vehicle can transport.}
\item{\code{consumables}}{The maximum length of time that this vehicle can provide consumables for its entire crew without having to resupply.}
\item{\code{hyperdrive_rating}}{The class of this starships hyperdrive.}
\item{\code{MGLT}}{The Maximum number of Megalights this starship can travel in a standard hour. A Megalight is a standard unit of distance and has never been defined before within the Star Wars universe. This figure is only really useful for measuring the difference in speed of starships. We can assume it is similar to AU, the distance between our Sun (Sol) and Earth.}
}
}
\usage{
vehicles
}
\description{
A Starship or vehicle in the Star Wars universe.
}
\references{
\url{https://swapi.dev/}
}
\keyword{datasets}
