% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people}
\alias{people}
\title{People}
\format{
A data frame with 82 rows and 10 variables:
\describe{
\item{\code{name}}{The name of this person.}
\item{\code{height}}{The height of this person in meters.}
\item{\code{mass}}{The mass of this person in kilograms.}
\item{\code{hair_color}}{The hair color of this person.}
\item{\code{skin_color}}{The skin color of this person.}
\item{\code{eye_color}}{The eye color of this person.}
\item{\code{birth_year}}{The birth year of this person. BBY (Before the Battle of Yavin) or ABY (After the Battle of Yavin).}
\item{\code{sex}}{The biological sex of the character. One of \code{male}, \code{female}, \code{hermaphroditic}, or \code{none}.}
\item{\code{gender}}{The gender role or gender identity of the character.}
\item{\code{homeworld}}{The planet the character was born on.}
\item{\code{species}}{The species of the character.}
}
}
\usage{
people
}
\description{
Characters within the Star Wars universe.
}
\references{
\url{https://swapi.dev/}
}
\keyword{datasets}
